Refactoring case select component
