+++ client/app/reader/constants.js
@@ -77,6 +77,7 @@ export const COMMENT_ACCORDION_KEY = 'Comments';
 // case select related constants
 export const RECEIVE_MULTIPLE_APPEALS_USING_VETERAN_ID = 'RECEIVE_MULTIPLE_APPEALS_USING_VETERAN_ID';
 export const RECEIVE_APPEALS_USING_VETERAN_ID_FAILED = 'RECEIVE_APPEALS_USING_VETERAN_ID_FAILED';
+export const RECEIVE_APPEALS_USING_VETERAN_ID = 'RECEIVE_APPEALS_USING_VETERAN_ID';
 export const CASE_SELECT_APPEAL = 'CASE_SELECT_APPEAL';
 export const SET_CASE_SELECT_SEARCH = 'SET_CASE_SELECT_SEARCH';
 export const CLEAR_SELECTED_APPEAL = 'CLEAR_SELECTED_APPEAL';
